+++ lib/Sdk/Buttons.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: must_be_immutable
+// ignore_for_file: must_be_immutable, file_names, camel_case_types, prefer_typing_uninitialized_variables, prefer_const_constructors
 
 import 'package:flutter/material.dart';