+++ lib/Sdk/Percent_indicator.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: must_be_immutable, dead_code
+// ignore_for_file: must_be_immutable, dead_code, file_names, camel_case_types, prefer_typing_uninitialized_variables, prefer_const_constructors, non_constant_identifier_names
 
 import 'package:flutter/material.dart';