+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+1. DSIBottomNav - alignment added
+2. DSI_BOTTOM_NAV_BUTTON - color added
+
 ## 0.1.0
 1. androidId
 2. H1